introduce-weight-disk-update-controller-move: Hand disk-based weight reload over to WeightDiskUpdateController
Pure physical move per MECH_COMMIT_SPLIT. Cut 4 @staticmethod methods
(``update_weights_from_disk``, ``_update_model_path_info``,
``_wait_for_model_update_from_disk``, ``_handle_update_weights_from_disk_req_output``)
from TokenizerManager and 1 @staticmethod method
(``_update_weight_version_if_provided``) from TokenizerControlMixin; paste
into WeightDiskUpdateController (drop @staticmethod, replace
``self: "WeightDiskUpdateController"`` → plain ``self``). Flip the
``__post_init__`` dispatcher entry from the lambda forwarder + late-TM-import
to a direct method reference. Caller prefix replacement:
``TokenizerManager.<method>(self.weight_disk_update_controller, ...)`` →
``self.weight_disk_update_controller.<method>(...)`` (TM + mixin sibling
methods + entrypoints); ditto for the
``TokenizerControlMixin._update_weight_version_if_provided`` call sites.

Refactor chain ID: introduce-weight-disk-update-controller-move
